Broke out canvas into separate component.
The main benefit of this (albeit at the cost of some complexity and
headaches) was to remove the responsibility from the driver for kicking
the app into gear. Instead, the child component just renders its input,
with a small bootstrapping step in the PsFrac view.

Could have done the same thing with a lifecycle component at the top
level, but it is nice to separate the "UI" from the rendering.